+++ PKHeX.Core/Legality/Encounters/EncounterTrade/EncounterTrade6.cs
@@ -4,14 +4,14 @@
     /// Generation 6 Trade Encounter
     /// </summary>
     /// <inheritdoc cref="EncounterTrade"/>
-    public sealed record EncounterTrade6(GameVersion Version, int OT_Memory, int OT_Intensity, int OT_Feeling, int OT_TextVar) : EncounterTrade(Version), IMemoryOT
+    public sealed record EncounterTrade6(GameVersion Version, byte OT_Memory, byte OT_Intensity, byte OT_Feeling, ushort OT_TextVar) : EncounterTrade(Version), IMemoryOT
     {
         public override int Generation => 6;
         public override int Location => Locations.LinkTrade6NPC;
-        public int OT_Memory { get; set; } = OT_Memory;
-        public int OT_Intensity { get; set; } = OT_Intensity;
-        public int OT_Feeling { get; set; } = OT_Feeling;
-        public int OT_TextVar { get; set; } = OT_TextVar;
+        public byte OT_Memory { get; set; } = OT_Memory;
+        public byte OT_Intensity { get; set; } = OT_Intensity;
+        public byte OT_Feeling { get; set; } = OT_Feeling;
+        public ushort OT_TextVar { get; set; } = OT_TextVar;
 
         protected override void ApplyDetails(ITrainerInfo sav, EncounterCriteria criteria, PKM pk)
         {
